+++ src/operators/tensor/implementations/tensor_complex64.cairo
@@ -128,23 +128,23 @@ impl Complex64Tensor of TensorTrait<complex64> {
         math::log::log(*self)
     }
 
-    fn equal(self: @Tensor<complex64>, other: @Tensor<complex64>) -> Tensor<usize> {
+    fn equal(self: @Tensor<complex64>, other: @Tensor<complex64>) -> Tensor<i32> {
         math::equal::equal(self, other)
     }
 
-    fn greater(self: @Tensor<complex64>, other: @Tensor<complex64>) -> Tensor<usize> {
+    fn greater(self: @Tensor<complex64>, other: @Tensor<complex64>) -> Tensor<i32> {
         panic(array!['not supported!'])
     }
 
-    fn greater_equal(self: @Tensor<complex64>, other: @Tensor<complex64>) -> Tensor<usize> {
+    fn greater_equal(self: @Tensor<complex64>, other: @Tensor<complex64>) -> Tensor<i32> {
         panic(array!['not supported!'])
     }
 
-    fn less(self: @Tensor<complex64>, other: @Tensor<complex64>) -> Tensor<usize> {
+    fn less(self: @Tensor<complex64>, other: @Tensor<complex64>) -> Tensor<i32> {
         panic(array!['not supported!'])
     }
 
-    fn less_equal(self: @Tensor<complex64>, other: @Tensor<complex64>) -> Tensor<usize> {
+    fn less_equal(self: @Tensor<complex64>, other: @Tensor<complex64>) -> Tensor<i32> {
         panic(array!['not supported!'])
     }
 
@@ -206,11 +206,11 @@ impl Complex64Tensor of TensorTrait<complex64> {
         math::atan::atan(*self)
     }
 
-    fn xor(self: @Tensor<complex64>, other: @Tensor<complex64>) -> Tensor<usize> {
+    fn xor(self: @Tensor<complex64>, other: @Tensor<complex64>) -> Tensor<i32> {
         panic(array!['not supported!'])
     }
 
-    fn or(self: @Tensor<complex64>, other: @Tensor<complex64>) -> Tensor<usize> {
+    fn or(self: @Tensor<complex64>, other: @Tensor<complex64>) -> Tensor<i32> {
         panic(array!['not supported!'])
     }
 
@@ -347,7 +347,7 @@ impl Complex64Tensor of TensorTrait<complex64> {
         panic(array!['not supported!'])
     }
 
-    fn and(self: @Tensor<bool>, other: @Tensor<bool>) -> Tensor<usize> {
+    fn and(self: @Tensor<bool>, other: @Tensor<bool>) -> Tensor<i32> {
         math::and::and(self, other)
     }
