+++ src/operators/tensor/implementations/tensor_i32.cairo
@@ -117,23 +117,23 @@ impl I32Tensor of TensorTrait<i32> {
         panic(array!['not supported!'])
     }
 
-    fn equal(self: @Tensor<i32>, other: @Tensor<i32>) -> Tensor<usize> {
+    fn equal(self: @Tensor<i32>, other: @Tensor<i32>) -> Tensor<i32> {
         math::equal::equal(self, other)
     }
 
-    fn greater(self: @Tensor<i32>, other: @Tensor<i32>) -> Tensor<usize> {
+    fn greater(self: @Tensor<i32>, other: @Tensor<i32>) -> Tensor<i32> {
         math::greater::greater(self, other)
     }
 
-    fn greater_equal(self: @Tensor<i32>, other: @Tensor<i32>) -> Tensor<usize> {
+    fn greater_equal(self: @Tensor<i32>, other: @Tensor<i32>) -> Tensor<i32> {
         math::greater_equal::greater_equal(self, other)
     }
 
-    fn less(self: @Tensor<i32>, other: @Tensor<i32>) -> Tensor<usize> {
+    fn less(self: @Tensor<i32>, other: @Tensor<i32>) -> Tensor<i32> {
         math::less::less(self, other)
     }
 
-    fn less_equal(self: @Tensor<i32>, other: @Tensor<i32>) -> Tensor<usize> {
+    fn less_equal(self: @Tensor<i32>, other: @Tensor<i32>) -> Tensor<i32> {
         math::less_equal::less_equal(self, other)
     }
 
@@ -195,11 +195,11 @@ impl I32Tensor of TensorTrait<i32> {
         panic(array!['not supported!'])
     }
 
-    fn xor(self: @Tensor<i32>, other: @Tensor<i32>) -> Tensor<usize> {
+    fn xor(self: @Tensor<i32>, other: @Tensor<i32>) -> Tensor<i32> {
         math::xor::xor(self, other)
     }
 
-    fn or(self: @Tensor<i32>, other: @Tensor<i32>) -> Tensor<usize> {
+    fn or(self: @Tensor<i32>, other: @Tensor<i32>) -> Tensor<i32> {
         math::or::or(self, other)
     }
 
@@ -372,7 +372,7 @@ impl I32Tensor of TensorTrait<i32> {
         core_tensor::clip(self, min, max)
     }
 
-    fn and(self: @Tensor<bool>, other: @Tensor<bool>) -> Tensor<usize> {
+    fn and(self: @Tensor<bool>, other: @Tensor<bool>) -> Tensor<i32> {
         math::and::and(self, other)
     }
