+++ src/operators/tensor/implementations/tensor_i8.cairo
@@ -115,23 +115,23 @@ impl I8Tensor of TensorTrait<i8> {
         panic(array!['not supported!'])
     }
 
-    fn equal(self: @Tensor<i8>, other: @Tensor<i8>) -> Tensor<usize> {
+    fn equal(self: @Tensor<i8>, other: @Tensor<i8>) -> Tensor<i32> {
         math::equal::equal(self, other)
     }
 
-    fn greater(self: @Tensor<i8>, other: @Tensor<i8>) -> Tensor<usize> {
+    fn greater(self: @Tensor<i8>, other: @Tensor<i8>) -> Tensor<i32> {
         math::greater::greater(self, other)
     }
 
-    fn greater_equal(self: @Tensor<i8>, other: @Tensor<i8>) -> Tensor<usize> {
+    fn greater_equal(self: @Tensor<i8>, other: @Tensor<i8>) -> Tensor<i32> {
         math::greater_equal::greater_equal(self, other)
     }
 
-    fn less(self: @Tensor<i8>, other: @Tensor<i8>) -> Tensor<usize> {
+    fn less(self: @Tensor<i8>, other: @Tensor<i8>) -> Tensor<i32> {
         math::less::less(self, other)
     }
 
-    fn less_equal(self: @Tensor<i8>, other: @Tensor<i8>) -> Tensor<usize> {
+    fn less_equal(self: @Tensor<i8>, other: @Tensor<i8>) -> Tensor<i32> {
         math::less_equal::less_equal(self, other)
     }
 
@@ -193,11 +193,11 @@ impl I8Tensor of TensorTrait<i8> {
         panic(array!['not supported!'])
     }
 
-    fn xor(self: @Tensor<i8>, other: @Tensor<i8>) -> Tensor<usize> {
+    fn xor(self: @Tensor<i8>, other: @Tensor<i8>) -> Tensor<i32> {
         math::xor::xor(self, other)
     }
 
-    fn or(self: @Tensor<i8>, other: @Tensor<i8>) -> Tensor<usize> {
+    fn or(self: @Tensor<i8>, other: @Tensor<i8>) -> Tensor<i32> {
         math::or::or(self, other)
     }
 
@@ -376,7 +376,7 @@ impl I8Tensor of TensorTrait<i8> {
         core_tensor::clip(self, min, max)
     }
 
-    fn and(self: @Tensor<bool>, other: @Tensor<bool>) -> Tensor<usize> {
+    fn and(self: @Tensor<bool>, other: @Tensor<bool>) -> Tensor<i32> {
         math::and::and(self, other)
     }
