+++ src/operators/tensor/implementations/tensor_u32.cairo
@@ -114,23 +114,23 @@ impl U32Tensor of TensorTrait<u32> {
         panic(array!['not supported!'])
     }
 
-    fn equal(self: @Tensor<u32>, other: @Tensor<u32>) -> Tensor<usize> {
+    fn equal(self: @Tensor<u32>, other: @Tensor<u32>) -> Tensor<i32> {
         math::equal::equal(self, other)
     }
 
-    fn greater(self: @Tensor<u32>, other: @Tensor<u32>) -> Tensor<usize> {
+    fn greater(self: @Tensor<u32>, other: @Tensor<u32>) -> Tensor<i32> {
         math::greater::greater(self, other)
     }
 
-    fn greater_equal(self: @Tensor<u32>, other: @Tensor<u32>) -> Tensor<usize> {
+    fn greater_equal(self: @Tensor<u32>, other: @Tensor<u32>) -> Tensor<i32> {
         math::greater_equal::greater_equal(self, other)
     }
 
-    fn less(self: @Tensor<u32>, other: @Tensor<u32>) -> Tensor<usize> {
+    fn less(self: @Tensor<u32>, other: @Tensor<u32>) -> Tensor<i32> {
         math::less::less(self, other)
     }
 
-    fn less_equal(self: @Tensor<u32>, other: @Tensor<u32>) -> Tensor<usize> {
+    fn less_equal(self: @Tensor<u32>, other: @Tensor<u32>) -> Tensor<i32> {
         math::less_equal::less_equal(self, other)
     }
 
@@ -192,11 +192,11 @@ impl U32Tensor of TensorTrait<u32> {
         panic(array!['not supported!'])
     }
 
-    fn xor(self: @Tensor<u32>, other: @Tensor<u32>) -> Tensor<usize> {
+    fn xor(self: @Tensor<u32>, other: @Tensor<u32>) -> Tensor<i32> {
         math::xor::xor(self, other)
     }
 
-    fn or(self: @Tensor<u32>, other: @Tensor<u32>) -> Tensor<usize> {
+    fn or(self: @Tensor<u32>, other: @Tensor<u32>) -> Tensor<i32> {
         math::or::or(self, other)
     }
 
@@ -320,7 +320,7 @@ impl U32Tensor of TensorTrait<u32> {
         core_tensor::clip(self, min, max)
     }
 
-    fn and(self: @Tensor<bool>, other: @Tensor<bool>) -> Tensor<usize> {
+    fn and(self: @Tensor<bool>, other: @Tensor<bool>) -> Tensor<i32> {
         math::and::and(self, other)
     }
